refactor: renamed slices to include 'market'
